Move backend API stubs to their own module
