--- app/src/main/java/com/psu/capstonew17/backend/api/TestStub.java
@@ -1,14 +1,0 @@
-package com.psu.capstonew17.backend.api;
-
-import java.util.Iterator;
-
-/**
- * Created by Tim on 4/27/2017.
- */
-
-public class TestStub implements Test{
-    @Override
-    public Statistics getStats() {
-        return null;
-    }
-}
