+++ app/src/main/java/com/psu/capstonew17/backend/api/stubs/SharingManagerStub.java
@@ -1,6 +1,13 @@
-package com.psu.capstonew17.backend.api;
+package com.psu.capstonew17.backend.api.stubs;
 
 import android.graphics.Bitmap;
+
+import com.psu.capstonew17.backend.api.Card;
+import com.psu.capstonew17.backend.api.Deck;
+import com.psu.capstonew17.backend.api.SharingManager;
+import com.psu.capstonew17.backend.api.SharingReceiveListener;
+import com.psu.capstonew17.backend.api.SharingTransmitListener;
+
 import java.util.List;
 
 /**
