+++ app/src/main/java/com/psu/capstonew17/backend/api/stubs/SharingReceiveListenerStub.java
@@ -1,4 +1,6 @@
-package com.psu.capstonew17.backend.api;
+package com.psu.capstonew17.backend.api.stubs;
+
+import com.psu.capstonew17.backend.api.SharingReceiveListener;
 
 /**
  * Created by Tim on 4/27/2017.
