+++ app/src/main/java/com/psu/capstonew17/backend/api/stubs/TestManagerStub.java
@@ -1,4 +1,8 @@
-package com.psu.capstonew17.backend.api;
+package com.psu.capstonew17.backend.api.stubs;
+import com.psu.capstonew17.backend.api.Deck;
+import com.psu.capstonew17.backend.api.Test;
+import com.psu.capstonew17.backend.api.TestManager;
+
 import java.util.List;
 /**
  * Created by Tim on 4/27/2017.
