+++ app/src/main/java/com/psu/capstonew17/backend/api/stubs/VideoManagerStub.java
@@ -1,4 +1,6 @@
-package com.psu.capstonew17.backend.api;
+package com.psu.capstonew17.backend.api.stubs;
+
+import com.psu.capstonew17.backend.api.VideoManager;
 
 import java.io.File;
