+++ app/src/main/java/com/psu/capstonew17/backend/api/stubs/VideoStub.java
@@ -1,12 +1,14 @@
-package com.psu.capstonew17.backend.api;
+package com.psu.capstonew17.backend.api.stubs;
 
 import android.media.MediaPlayer;
+
+import com.psu.capstonew17.backend.api.Video;
 
 /**
  * Created by Tim on 4/21/2017.
  */
 
-public class VideoStub implements Video{
+public class VideoStub implements Video {
 
     public void configurePlayer(MediaPlayer player){
         return;
